(FM-6472) Add logic to determine facter path
Prior to this commit, facter_task ran "facter" in the existing
PATH.  This will cause unexpected failures if facter happens
to not be on the path, isn't present on the system, or there
happens to be another program called "facter" on the PATH.  With
this commit, we now set reasonable default values for the facter
path on both Windows and Linux.  In addition, we attempt to look
for the specific path set on Windows installations by the puppet
environment.