(MODULES-10602) workaround for "765: unexpected token at ''" error
When any fact on the target uses `Puppet::Util::Execution.execute`,
something between facter and this task gets confused and returns a NUL
character on stdout, instead of the JSON. This change patches the task
to use Puppet's execution method to workaround the symptom.

Note that this now completely discards facter's stderr making debugging
fact failures even harder. Alternatively, we could remove `combine: false`,
which would combine the stdout and stderr into a single stream, which
in turn would have any fact resolution error cause the task unusable
because the result is not parsable JSON anymore.

See also FACT-2682 for a more detailed write-up of the underlying issue.